【代码整理】C 风格类型转换修改为 C++ 风格

+++ Lucky/Source/Lucky/Core/UUID.h
@@ -30,7 +30,7 @@ namespace std
     {
         std::size_t operator()(const Lucky::UUID& uuid) const
         {
-            return hash<uint64_t>()((uint64_t)uuid);
+            return hash<uint64_t>()(uuid);
         }
     };
 }
